chore(cleanup): remove dead pre-#413 supervisor files (#438)
Delete 4 dead production files (supervisor.py, supervisor_state.py,
tools.py, manager_tools.py) and 2 dead test files from the old
supervisor architecture replaced by create_agent SDK in #413.

Inline test-only functions (build_tools_for_role,
create_manager_nurturing_tools, create_crm_score_sync_tool) into
the 3 test files that depended on them. Update agents/__init__.py
docstring to reference #413.
